pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project]
name = "bgg-dash-viewer"
version = "0.1.0"
description = "A Dash-based viewer for BoardGameGeek game data"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dash>=2.15.0",
    "dash-bootstrap-components>=1.5.0",
    "plotly>=5.18.0",
    "pandas>=2.1.0",
    "google-cloud-bigquery>=3.14.0",
    "pyarrow>=15.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.1",
    "pandas-gbq>=0.29.1",
    "google-cloud-bigquery-storage>=2.32.0",
    "google-auth>=2.40.3",
    "flask-caching>=2.1.0",
    "gunicorn>=21.2.0", # For production deployment
]

[project.optional-dependencies]
dev = [
    "black>=24.1.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
    "pytest>=8.0.0",
]

[project.scripts]
bgg-dash-viewer = "src.app:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"
pythonpath = ["src"]

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.ruff]
line-length = 100
target-version = "py312"
select = ["E", "F", "B", "I", "N", "UP", "PL", "RUF"]
